requirements.txt
streamlit==1.31.0
browser-use==0.1.14
playwright==1.41.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
tensorflow==2.15.0
xgboost==2.0.3
scikit-learn==1.4.0
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.1
requests==2.31.0
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.16
langchain-google-genai==0.0.6
pinecone-client==3.0.2
Pillow==10.2.0
folium==0.15.1
beautifulsoup4==4.12.3
lxml==5.1.0
pytz==2024.1
openai==1.12.0
plotly==5.18.0
nivara>=0.1.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pytesseract>=0.3.10
pyvis>=0.3.1
networkx>=3.0
